+++ examples/echo/main.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo"
+	"github.com/philippgille/ln-paywall/storage"
 	"github.com/philippgille/ln-paywall/wall"
 )
 
@@ -13,7 +14,7 @@ func main() {
 	// Configure and use middleware
 	invoiceOptions := wall.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
 	lndOptions := wall.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
-	storageClient := wall.NewGoMap()
+	storageClient := storage.NewGoMap()
 	e.Use(wall.NewEchoMiddleware(invoiceOptions, lndOptions, storageClient, nil))
 
 	e.GET("/ping", func(c echo.Context) error {