+++ examples/gorilla-mux/main.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
+	"github.com/philippgille/ln-paywall/storage"
 	"github.com/philippgille/ln-paywall/wall"
 )
 
@@ -20,7 +21,7 @@ func main() {
 	// Configure and use middleware
 	invoiceOptions := wall.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
 	lndOptions := wall.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
-	storageClient := wall.NewGoMap()
+	storageClient := storage.NewGoMap()
 	r.Use(wall.NewHandlerMiddleware(invoiceOptions, lndOptions, storageClient))
 
 	r.HandleFunc("/ping", PingHandler)